added image to design ch
