+++ src/components/dom/DynamicExperienceMasks.tsx
@@ -1,0 +1,11 @@
+export default function DynamicExperienceMasks() {
+  // Adjusts canvas position by adjusting surrounding masks
+  return (
+    <>
+      <div className='mask is-top' />
+      <div className='mask is-right' />
+      <div className='mask is-bottom' />
+      <div className='mask is-left' />
+    </>
+  )
+}
